project AI (1)/project AI/project AI/requirements.txt
neo4j==5.14.1
spacy==3.7.2
pgmpy==0.1.24
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0 